Activate symbol layer with shift tap
Too much thumb movement on left hand. Use tap dance to trigger
the symbol layer when tapping the shift key, and regular
shift when held. Still dialing in the tapping term; it's at
a point where I can activate the sym layer and do a pair of
braces, but I still need to be consciously patient when I want
to type capitals.

I've used the indicator lights to display the status of the tap
function. Can't get the RHS indicator to work though, even though
it works fine in an RGB layer.

The old symbol layer key now activates just arrow keys. Might
find other things to add to this layer. I could fit the arrows
in the fn layer, but I want to keep that for meta/system actions,
and avoid adding keys that get typed absent-mindedly (we have
power down on that layer).