Adjust sym and qwerty layer
Add non-alpha keys to "quick shift" sym layer to speed up rhythm of
typing symbols. If we don't wait long enough for shift to be held
we still get the symbol we want. Only need to press long enough for
alphas.

Re-add grave and get rid of grave escape - too many instances of the
win + esc = grave not working, and also annoying having to hit the
win key to get a grave. Move delete, home, end to match keycap layout.